fix tests on store
